Fix re-signaling daemon on reload

+++ core/daemons.c
@@ -235,13 +235,17 @@ void uwsgi_detach_daemons() {
 			// try to stop daemon gracefully, kill it if it won't die
 			// if mercy is not set then wait up to 3 seconds
 			time_t timeout = uwsgi_now() + (uwsgi.reload_mercy ? uwsgi.reload_mercy : 3);
+			int has_been_signalled = 0;
 			int waitpid_status;
 			while (!kill(ud->pid, 0)) {
-				if (uwsgi_instance_is_reloading && ud->reload_signal > 0) {
-					kill(-(ud->pid), ud->reload_signal);
-				}
-				else {
-					kill(-(ud->pid), ud->stop_signal);
+				if (!has_been_signalled) {
+					if (uwsgi_instance_is_reloading && ud->reload_signal > 0) {
+						kill(-(ud->pid), ud->reload_signal);
+					}
+					else {
+						kill(-(ud->pid), ud->stop_signal);
+					}
+					has_been_signalled = 1;
 				}
 				sleep(1);
 				waitpid(ud->pid, &waitpid_status, WNOHANG);
@@ -258,10 +262,17 @@ void uwsgi_detach_daemons() {
 		// smart daemons that have to be notified when master is reloading or stopping
 		if (ud->notifypid && ud->pid > 0 && ud->pidfile) {
 			if (uwsgi_instance_is_reloading) {
-				kill(-(ud->pid), ud->reload_signal > 0 ? ud->reload_signal : SIGHUP);
+				int signum = ud->reload_signal > 0 ? ud->reload_signal : SIGHUP;
+				if (kill(ud->pid, signum) < 0) {
+					uwsgi_log("[uwsgi-daemons] error notifying daemon of reload (pid: %d) (pidfile: %s) (signal: %d) errno: %d\n", (int) ud->pid, ud->pidfile, signum, errno);
+				}
 			}
 			else {
-				kill(-(ud->pid), ud->stop_signal);
+				if (kill(ud->pid, ud->stop_signal) < 0) {
+					uwsgi_log("[uwsgi-daemons] error notifying daemon of stop (pid: %d) (pidfile: %s) (signal: %d) errno: %d\n", (int) ud->pid, ud->pidfile, ud->stop_signal, errno);
+				}
+				// unregister daemon to prevent it from being respawned if it chose to die alongside the master
+				ud->registered = 0;
 			}
 		}
 		ud = ud->next;
